refactor fuzzy search into util
